removed old test file
